+++ ibmmq/cmqc_linux_s390x.go
@@ -30,8 +30,8 @@ package ibmmq
 ****************************************************************
 *
 *   <BEGIN_BUILDINFO>
-*   Generated on:  2/7/22 6:36 PM
-*   Build Level:   p925-L220207
+*   Generated on:  6/6/22 6:08 PM
+*   Build Level:   p930-L220606
 *   Build Type:    Production
 *   <END_BUILDINFO>
  */
@@ -617,6 +617,7 @@ const (
 	MQCA_ENV_DATA                  int32  = 2007
 	MQCA_FIRST                     int32  = 2001
 	MQCA_IGQ_USER_ID               int32  = 2041
+	MQCA_INITIAL_KEY               int32  = 2054
 	MQCA_INITIATION_Q_NAME         int32  = 2008
 	MQCA_INSTALLATION_DESC         int32  = 2115
 	MQCA_INSTALLATION_NAME         int32  = 2116
@@ -674,6 +675,7 @@ const (
 	MQCA_SSL_KEY_LIBRARY           int32  = 2069
 	MQCA_SSL_KEY_MEMBER            int32  = 2070
 	MQCA_SSL_KEY_REPOSITORY        int32  = 2049
+	MQCA_SSL_KEY_REPO_PASSWORD     int32  = 2055
 	MQCA_STDERR_DESTINATION        int32  = 2084
 	MQCA_STDOUT_DESTINATION        int32  = 2083
 	MQCA_STORAGE_CLASS             int32  = 2022
@@ -1019,7 +1021,7 @@ const (
 	MQCMDI_SEC_SIGNOFF_ERROR       int32  = 17
 	MQCMDI_SEC_TIMER_ZERO          int32  = 14
 	MQCMDI_SEC_UPPERCASE           int32  = 21
-	MQCMDL_CURRENT_LEVEL           int32  = 925
+	MQCMDL_CURRENT_LEVEL           int32  = 930
 	MQCMDL_LEVEL_1                 int32  = 100
 	MQCMDL_LEVEL_101               int32  = 101
 	MQCMDL_LEVEL_110               int32  = 110
@@ -1066,6 +1068,7 @@ const (
 	MQCMDL_LEVEL_923               int32  = 923
 	MQCMDL_LEVEL_924               int32  = 924
 	MQCMDL_LEVEL_925               int32  = 925
+	MQCMDL_LEVEL_930               int32  = 930
 	MQCMD_ACCOUNTING_MQI           int32  = 167
 	MQCMD_ACCOUNTING_Q             int32  = 168
 	MQCMD_ACTIVITY_MSG             int32  = 69
@@ -1355,10 +1358,12 @@ const (
 	MQCRC_TRANSID_NOT_AVAILABLE    int32  = 9
 	MQCSP_AUTH_NONE                int32  = 0
 	MQCSP_AUTH_USER_ID_AND_PWD     int32  = 1
-	MQCSP_CURRENT_LENGTH           int32  = 56
-	MQCSP_CURRENT_VERSION          int32  = 1
+	MQCSP_CURRENT_LENGTH           int32  = 80
+	MQCSP_CURRENT_VERSION          int32  = 2
 	MQCSP_LENGTH_1                 int32  = 56
+	MQCSP_LENGTH_2                 int32  = 80
 	MQCSP_VERSION_1                int32  = 1
+	MQCSP_VERSION_2                int32  = 2
 	MQCSRV_CONVERT_NO              int32  = 0
 	MQCSRV_CONVERT_YES             int32  = 1
 	MQCSRV_DLQ_NO                  int32  = 0
@@ -1562,6 +1567,7 @@ const (
 	MQFB_COD                       int32  = 260
 	MQFB_DATA_LENGTH_NEGATIVE      int32  = 292
 	MQFB_DATA_LENGTH_TOO_BIG       int32  = 293
+	MQFB_DATA_LENGTH_TOO_SHORT     int32  = 299
 	MQFB_DATA_LENGTH_ZERO          int32  = 291
 	MQFB_EXPIRATION                int32  = 258
 	MQFB_IIH_ERROR                 int32  = 296
@@ -4272,13 +4278,14 @@ const (
 	MQSCOPE_AS_PARENT              int32  = 1
 	MQSCOPE_QMGR                   int32  = 4
 	MQSCO_CELL                     int32  = 2
-	MQSCO_CURRENT_LENGTH           int32  = 632
-	MQSCO_CURRENT_VERSION          int32  = 5
+	MQSCO_CURRENT_LENGTH           int32  = 648
+	MQSCO_CURRENT_VERSION          int32  = 6
 	MQSCO_LENGTH_1                 int32  = 536
 	MQSCO_LENGTH_2                 int32  = 544
 	MQSCO_LENGTH_3                 int32  = 560
 	MQSCO_LENGTH_4                 int32  = 568
 	MQSCO_LENGTH_5                 int32  = 632
+	MQSCO_LENGTH_6                 int32  = 648
 	MQSCO_Q_MGR                    int32  = 1
 	MQSCO_RESET_COUNT_DEFAULT      int32  = 0
 	MQSCO_VERSION_1                int32  = 1
@@ -4286,6 +4293,7 @@ const (
 	MQSCO_VERSION_3                int32  = 3
 	MQSCO_VERSION_4                int32  = 4
 	MQSCO_VERSION_5                int32  = 5
+	MQSCO_VERSION_6                int32  = 6
 	MQSCYC_MIXED                   int32  = 1
 	MQSCYC_UPPER                   int32  = 0
 	MQSD_CURRENT_LENGTH            int32  = 344
@@ -4942,6 +4950,7 @@ const (
 	MQ_FUNCTION_LENGTH             int32  = 4
 	MQ_GROUP_ADDRESS_LENGTH        int32  = 264
 	MQ_GROUP_ID_LENGTH             int32  = 24
+	MQ_INITIAL_KEY_LENGTH          int32  = 256
 	MQ_INSTALLATION_DESC_LENGTH    int32  = 64
 	MQ_INSTALLATION_NAME_LENGTH    int32  = 16
 	MQ_INSTALLATION_PATH_LENGTH    int32  = 256
@@ -5025,11 +5034,13 @@ const (
 	MQ_SSL_CIPHER_SPEC_LENGTH      int32  = 32
 	MQ_SSL_CIPHER_SUITE_LENGTH     int32  = 32
 	MQ_SSL_CRYPTO_HARDWARE_LENGTH  int32  = 256
+	MQ_SSL_ENCRYP_KEY_REPO_PWD_LEN int32  = 1536
 	MQ_SSL_HANDSHAKE_STAGE_LENGTH  int32  = 32
 	MQ_SSL_KEY_LIBRARY_LENGTH      int32  = 44
 	MQ_SSL_KEY_MEMBER_LENGTH       int32  = 8
 	MQ_SSL_KEY_PASSPHRASE_LENGTH   int32  = 1024
 	MQ_SSL_KEY_REPOSITORY_LENGTH   int32  = 256
+	MQ_SSL_KEY_REPO_PWD_LEN        int32  = 1024
 	MQ_SSL_PEER_NAME_LENGTH        int32  = 1024
 	MQ_SSL_SHORT_PEER_NAME_LENGTH  int32  = 256
 	MQ_START_CODE_LENGTH           int32  = 4