+++ mqmetric/globals.go
@@ -51,6 +51,8 @@ type connectionInfo struct {
 	useStatus            bool
 	useResetQStats       bool
 	showInactiveChannels bool
+	hideSvrConnJobname   bool
+	durableSubPrefix     string
 
 	// Only issue the warning about a '/' in an object name once.
 	globalSlashWarning bool
@@ -129,6 +131,7 @@ func newConnectionInfo(key string) *connectionInfo {
 	ci.useStatus = false
 	ci.useResetQStats = false
 	ci.showInactiveChannels = false
+	ci.hideSvrConnJobname = false
 
 	ci.globalSlashWarning = false
 	ci.localSlashWarning = false