+++ mqmetric/status.go
@@ -128,11 +128,13 @@ func statusTimeDiff(now time.Time, d string, t string) int64 {
 		if err == nil {
 			diff := now.Sub(parsedT).Seconds() + ci.tzOffsetSecs
 
-			if diff < -(60 * 5) { // Cannot have status from the future but allow a tiny amount of flex
-				if !timeTravelWarningIssued {
-					logError("Status reports appear to be from the future. Difference is approximately %d seconds. Check the TZ Offset value in the program configuration.", int64(-diff))
-					//logError("statusTimeDiff d:%s t:%s diff:%f tzoffset: %f err:%v\n", d, t, diff, ci.tzOffsetSecs, err)
-					timeTravelWarningIssued = true
+			if diff < 0 {
+				if diff < -(60 * 5) { // Cannot have status from the future but allow a tiny amount of flex
+					if !timeTravelWarningIssued {
+						logError("Status reports appear to be from the future. Difference is approximately %d seconds. Check the TZ Offset value in the program configuration.", int64(-diff))
+						logDebug("statusTimeDiff d:%s t:%s diff:%f tzoffset: %f err:%v\n", d, t, diff, ci.tzOffsetSecs, err)
+						timeTravelWarningIssued = true
+					}
 				}
 				diff = 0
 			}